# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "North Bound Interface (NBI) REST API exposed by voltha-go-controller (VGC).\nIt provides configuration and operational endpoints for subscribers, bandwidth\nprofiles, IGMP proxy, multicast, flows, devices, ports, meters, groups, DHCP\nallocations, MAC learning and the OLT flow service.",
        "title": "VOLTHA Go Controller NBI REST API",
        "contact": {
            "name": "Open Networking Foundation",
            "url": "https://opennetworking.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8181",
    "basePath": "/vgc/v1",
    "paths": {
...
        "/meters/{id}": {
            "get": {
                "description": "Retrieve a specific meter by its identifier.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meters"
                ],
                "summary": "Get meter by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Meter identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/onosnbi.MeterList"
                        }
                    },
                    "404": {
                        "description": "Meter not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "onosnbi.Bands": {
            "type": "object",
            "properties": {
                "burstSize": {
                    "type": "integer"
                },
                "bytes": {
                    "type": "integer"
                },
                "packets": {
                    "type": "integer"
                },
                "prec": {
                    "type": "integer"
                },
                "rate": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "onosnbi.MeterList": {
            "type": "object",
            "properties": {
                "meters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/onosnbi.Meters"
                    }
                }
            }
        },
        "onosnbi.Meters": {
            "type": "object",
            "properties": {
                "appId": {
                    "type": "string"
                },
                "bands": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/onosnbi.Bands"
                    }
                },
                "burst": {
                    "type": "boolean"
                },
                "bytes": {
                    "type": "integer"
                },
                "deviceId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "life": {
                    "type": "integer"
                },
                "packets": {
                    "type": "integer"
                },
                "referenceCount": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
    }
}